+++ example/test/widget_test.dart
@@ -13,15 +13,10 @@ import 'package:securely_example/main.dart';
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    // wrap MyApp with a MaterialApp just like the real app does
+    await tester.pumpWidget(const MaterialApp(home: MyApp()));
 
-    // Verify that platform version is retrieved.
-    expect(
-      find.byWidgetPredicate(
-        (Widget widget) => widget is Text &&
-                           widget.data!.startsWith('Running on:'),
-      ),
-      findsOneWidget,
-    );
+    // Verify that one of the feature buttons is present.
+    expect(find.text('Check Debugger Status'), findsOneWidget);
   });
 }